Boiler plate react-app cleaned up
Cleaned up some built in commenting. Removed files included with template that are unecessary for my project.
Ran some tests to get a feel for what I was wanting structure wise. The remnants of those tests can be seen with a Nav component and a CircularIcon component currently lacking implementation.